+++ examples/hono/rpc.config.ts
@@ -1,4 +1,4 @@
-import { defineConfig } from '@thednp/rpc';
+import { defineConfig } from '@thednp/rpc/config';
 
 export default defineConfig({
   rpcPrefix: '_server',